# apps/server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  dict.createDictDataRequest:
    properties:
      cssClass:
        type: string
      dictLabel:
        type: string
      dictSort:
        type: integer
      dictValue:
        type: string
      isDefault:
        type: string
      listClass:
        type: string
      remark:
        type: string
      status:
        type: string
    required:
    - dictLabel
    - dictValue
    type: object
  dict.createDictTypeRequest:
    properties:
      dictName:
        type: string
      dictType:
        type: string
      remark:
        type: string
      status:
        type: string
    required:
    - dictName
    - dictType
    type: object
  dict.updateDictDataRequest:
    properties:
      cssClass:
        type: string
      dictLabel:
        type: string
      dictSort:
        type: integer
      dictValue:
        type: string
      isDefault:
        type: string
      listClass:
        type: string
      remark:
        type: string
      status:
        type: string
    type: object
  dict.updateDictTypeRequest:
    properties:
      dictName:
        type: string
      dictType:
        type: string
      remark:
        type: string
      status:
        type: string
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal admin platform API documentation.
  title: Admin Service API
  version: 0.1.0
paths:
...
  /v1/system/dicts:
    get:
      description: 按状态、名称、类型过滤字典类型
      parameters:
      - description: 字典状态
        in: query
        name: status
        type: string
      - description: 字典名称
        in: query
        name: dictName
        type: string
      - description: 字典类型
        in: query
        name: dictType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取字典类型列表
      tags:
      - System/Dict
    post:
      consumes:
      - application/json
      description: 创建字典类型
      parameters:
      - description: 字典类型参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dict.createDictTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 新增字典类型
      tags:
      - System/Dict
  /v1/system/dicts/{id}:
    delete:
      description: 根据ID删除字典类型
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除字典类型
      tags:
      - System/Dict
    get:
      description: 根据ID查询字典类型
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取字典类型详情
      tags:
      - System/Dict
    put:
      consumes:
      - application/json
      description: 更新字典类型信息
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 字典类型参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dict.updateDictTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 修改字典类型
      tags:
      - System/Dict
  /v1/system/dicts/{id}/data:
    get:
      description: 查询指定字典下的数据项
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      - description: 数据标签
        in: query
        name: dictLabel
        type: string
      - description: 数据值
        in: query
        name: dictValue
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取字典数据列表
      tags:
      - System/Dict
    post:
      consumes:
      - application/json
      description: 在字典下新增数据项
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 字典数据参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dict.createDictDataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 新增字典数据
      tags:
      - System/Dict
  /v1/system/dicts/{id}/data/{itemId}:
    delete:
      description: 删除指定字典下的数据项
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 数据ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除字典数据
      tags:
      - System/Dict
    put:
      consumes:
      - application/json
      description: 更新字典数据项
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 数据ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: 字典数据参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dict.updateDictDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 修改字典数据
      tags:
      - System/Dict
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"